Make properties of WebPixelAPI readonly, rename file
