+++ packages/web-pixels-extension/src/index.ts
@@ -1,2 +1,3 @@
 export {register} from './register';
 export type {ShopifyGlobal} from './globals';
+export * from './types';
