+++ packages/web-pixels-extension/src/types/index.ts
@@ -1,1 +1,1 @@
-export * from './runtime-api';
+export * from './web-pixel-api';
